// src/types/enums/common.emums.ts
export enum UserRoleEnum {
  ADMIN = 'admin',
  MANAGER = 'manager',
  USER = 'user',
}

export enum SubjectEnum {
  USER = 'user',
  ROLE = 'role',
  PERMISSION = 'permission',
  CONTACT_US = 'contact-us',
  PRIVACY_POLICY = 'privacy-policy',
  TERMS_AND_CONDITION = 'terms-and-condition',
  SEO = 'seo',
  PAYMENTS = 'payments',
  CATEGORY = 'category',
  PRODUCT = 'product',
  FAQ = 'faq',
  ORDER = 'order',

  ALL = 'all',
}

export enum ActionEnum {
  CREATE = 'create',
  READ = 'read',
  UPDATE = 'update',
  DELETE = 'delete',
}